+++ packages/project/src/Project.ts
@@ -15,6 +15,7 @@ import { Workspace } from './Workspace';
 import { buildConfig, extractConfig } from './util/config';
 import { Provisioner } from '@openfn/lexicon/lightning';
 import { SandboxMeta, UUID, WorkspaceConfig } from '@openfn/lexicon';
+import { parse } from './util/get-credential-name';
 
 const maybeCreateWorkflow = (wf: any) =>
   wf instanceof Workflow ? wf : new Workflow(wf);
@@ -32,6 +33,14 @@ type CLIMeta = {
   version?: number;
   alias?: string;
   forked_from?: Record<string, string>;
+};
+
+export type Credential = {
+  uuid: string;
+  name: string;
+
+  // TODO having the owner in the credential is controvertial and we may need to rethink this later
+  owner: string;
 };
 
 export class Project {
@@ -70,7 +79,7 @@ export class Project {
 
   collections: any;
 
-  credentials: string[];
+  credentials: Credential[];
 
   sandbox?: SandboxMeta;
 
@@ -239,6 +248,26 @@ export class Project {
       };
     }
     return result;
+  }
+
+  /**
+   * Find all project credentials referenced in all
+   * workflows and return it
+   */
+  buildCredentialMap(): Credential[] {
+    const creds: any = {};
+    for (const wf of this.workflows) {
+      for (const step of wf.steps) {
+        if (
+          typeof step.configuration === 'string' &&
+          !creds[step.configuration]
+        ) {
+          const { name, owner } = parse(step.configuration);
+          creds[step.configuration] = { name, owner };
+        }
+      }
+    }
+    return Object.values(creds);
   }
 
   // Compare this project with another and return a list of workflow changes