+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @openfn/cli
 
+## 1.28.0
+
+### Minor Changes
+
+- 1bbc8c4: Update credentials to use credential id, not UUID. This enables credentials to sync better with app projects.
+
+  WARNING: existing credential maps will break after pulling after this change. Update your credential maps to index on the new id values rather than the UUIDs.
+
+### Patch Changes
+
+- Updated dependencies [1bbc8c4]
+  - @openfn/project@0.14.0
+
 ## 1.27.0
 
 ### Minor Changes
@@ -8,7 +21,7 @@
 
 ### Patch Changes
 
-- 5d6237d: DEfault endpoint to app.openfn.org and improve error message
+- 5d6237d: Default endpoint to app.openfn.org and improve error message
 - Updated dependencies [92c0b49]
 - Updated dependencies [dd88099]
   - @openfn/project@0.13.1